+++ myokit/tests/test_simulation_cvodes.py
@@ -230,13 +230,45 @@ class SimulationTest(unittest.TestCase):
     def test_no_protocol(self):
         # Test running without a protocol.
 
+        # Check if pace was set to zero (see technical notes).
         self.sim.reset()
         self.sim.pre(50)
         self.sim.set_protocol(None)
-        d = self.sim.run(50).npview()
+        d = self.sim.run(50, log=['engine.pace']).npview()
+        self.assertTrue(np.all(d['engine.pace'] == 0))
 
-        # Check if pace was set to zero (see prop 651 / technical docs).
-        self.assertTrue(np.all(d['engine.pace'] == 0.0))
+        # Defined at the start? Then still counts as bound
+        self.assertRaisesRegex(
+            ValueError, 'not a literal',
+            self.sim.set_constant, 'engine.pace', 1)
+
+        # Check that pace is reset to zero when protocol is removed
+        x = 0.01    # Note: Must be low to stop sim crashing :D
+        self.sim.set_protocol(
+            myokit.pacing.blocktrain(duration=1000, level=x, period=1000))
+        d = self.sim.run(5, log=['engine.pace']).npview()
+        self.assertTrue(np.all(d['engine.pace'] == x))
+        self.sim.set_protocol(None)
+        d = self.sim.run(5, log=['engine.pace']).npview()
+        self.assertTrue(np.all(d['engine.pace'] == 0))
+        self.sim.set_protocol(self.protocol)
+
+        # No protocol set from the start? Then "pace" is set anyway for
+        # backwards compatibility
+        m = self.model.clone()
+        m.get('engine.pace').set_rhs(x)
+        s = myokit.Simulation(m)
+        d = s.run(50).npview()
+        self.assertIn('engine.pace', d.keys())
+        self.assertTrue(np.all(d['engine.pace'] == 0))
+
+        # But if user explicitly says no labels, then pace isn't treated in a
+        # special way.
+        s = myokit.Simulation(m, {})
+        self.assertRaises(ValueError, s.set_protocol, self.protocol)
+        d = s.run(50).npview()
+        self.assertNotIn('engine.pace', d.keys())
+        #self.assertTrue(np.all(d['engine.pace'] == x))  constant is not logged
 
     def test_wrong_label_set_pacing(self):
         # Test set_pacing with incorrect label